fix non numeric isssue

+++ app/Console/Commands/ProcessSupplierPaymentStatement.php
@@ -106,7 +106,7 @@ class ProcessSupplierPaymentStatement extends Command
                                     'tds' => $tds_amount,
                                     'tcs' => $tcs_amount,
                                     'adjustment_amount' => OrderPaymentDistribution::DEFAULT_ADJUSTMENT_AMOUNT,
-                                    'disburse_amount' => $order->total_amount - ($tds_amount + $tcs_amount + $processing_charges + $payment_gateway_charges + $refund_amount + OrderPaymentDistribution::DEFAULT_ADJUSTMENT_AMOUNT),
+                                    'disburse_amount' => (float)  $order->total_amount - ((float) $tds_amount + (float) $tcs_amount + (float) $processing_charges + (float) $payment_gateway_charges + (float) $refund_amount + (float) OrderPaymentDistribution::DEFAULT_ADJUSTMENT_AMOUNT),
                                     'payment_status' => $payment_status,
                                     'statement_date' => $payment_week,
                                     'payment_method' => SupplierPayment::PAYMENT_METHOD_BANK_TRANSFER
@@ -132,7 +132,7 @@ class ProcessSupplierPaymentStatement extends Command
                                 'tds' => $tds_amount,
                                 'tcs' => $tcs_amount,
                                 'adjustment_amount' => OrderPaymentDistribution::DEFAULT_ADJUSTMENT_AMOUNT,
-                                'disburse_amount' => $order->total_amount - ($tds_amount + $tcs_amount + $processing_charges + $payment_gateway_charges + $refund_amount + OrderPaymentDistribution::DEFAULT_ADJUSTMENT_AMOUNT),
+                                'disburse_amount' => (float)  $order->total_amount - ((float) $tds_amount + (float) $tcs_amount + (float) $processing_charges + (float) $payment_gateway_charges + (float) $refund_amount + (float) OrderPaymentDistribution::DEFAULT_ADJUSTMENT_AMOUNT),
                                 'payment_status' => $payment_status,
                                 'statement_date' => $payment_week,
                                 'payment_method' => SupplierPayment::PAYMENT_METHOD_BANK_TRANSFER
